+++ lib/src/config_builder/config_builder.dart
@@ -70,9 +70,10 @@ class ConfigBuilder {
 
   /// Creates a raw unused files config from given [excludePatterns].
   static UnusedFilesConfig getUnusedFilesConfigFromArgs(
-    Iterable<String> excludePatterns,
-  ) =>
-      UnusedFilesConfig.fromArgs(excludePatterns);
+    Iterable<String> excludePatterns, {
+    required bool isMonorepo,
+  }) =>
+      UnusedFilesConfig.fromArgs(excludePatterns, isMonorepo: isMonorepo);
 
   /// Creates a raw unused files config from given [options].
   static UnusedFilesConfig getUnusedFilesConfigFromOption(