+++ test/src/analyzers/unused_files_analyzer/unused_files_analyzer_test.dart
@@ -51,4 +51,5 @@ UnusedFilesConfig _createConfig({
     UnusedFilesConfig(
       excludePatterns: const [],
       analyzerExcludePatterns: analyzerExcludePatterns,
+      isMonorepo: false,
     );